requirements.txt
apns2==0.3.0
asn1crypto==0.24.0
beautifulsoup4==4.6.0
bleach==2.1.3
certifi==2018.1.18
cffi==1.11.5
chardet==3.0.4
click==6.7
coreapi==2.3.3
coreschema==0.0.4
cryptography==2.2.2
defusedxml==0.5.0
dj-database-url==0.5.0
Django==1.11
django-appconf==1.0.2
django-classy-tags==0.8.0
django-cms==3.5.2
django-contrib-comments==1.8.0
django-filter==1.1.0
django-formtools==2.1
django-haystack==2.8.1
django-leaflet==0.24.0
django-model-utils==3.1.1
django-polymorphic==2.0.2
django-sekizai==0.10.0
django-templated-mail==1.1.1
django-treebeard==4.3
djangocms-admin-style==1.2.8
djangorestframework==3.7.0
djangorestframework-jwt==1.11.0
djangorestframework-yaml==1.0.3
-e git+https://github.com/sunscrapers/djoser.git@815092c04a29be58ac67a4c31499970778b4fb8b#egg=djoser
docutils==0.14
filebrowser-safe==0.4.7
future==0.16.0
grappelli-safe==0.4.7
h2==2.6.2
hpack==3.0.0
html5lib==1.0.1
http-ece==1.0.5
httpie==0.9.9
hyper==0.7.0
hyperframe==3.2.0
idna==2.6
ipaddress==1.0.22
itypes==1.1.0
Jinja2==2.10
jsonfield==2.0.2
lockfile==0.12.2
MarkupSafe==1.0
oauthlib==2.0.7
Pillow==5.1.0
pinax-notifications==5.0.3
psycopg2==2.7.4
psycopg2-binary==2.7.4
py-vapid==1.3.1
pycparser==2.18
Pygments==2.2.0
PyJWT==1.6.1
pyparsing==2.2.0
python-daemon==2.1.2
python-dateutil==2.7.3
python3-openid==3.1.0
pytz==2018.3
pywebpush==1.7.0
PyYAML==3.12
requests==2.18.4
requests-oauthlib==0.8.0
service==0.4.1
setproctitle==1.1.10
six==1.11.0
social-auth-app-django==2.1.0
social-auth-core==1.7.0
start==0.2
svgwrite==1.1.12
Tree==0.2.3
tzlocal==1.5.1
Unidecode==1.0.22
uritemplate==3.0.0
urllib3==1.22
webencodings==0.5.1
